a kaggle contest
On E-commerce sites, the quality of product listing is crucial for improving search relevance and gaining customer attention. In this competition, you are provided a set of product names, description, and attributes, as well as the quality score of their listing as rated by real customers. You are challenged to build a model to automatically predict the quality of another set of product listings.

https://www.kaggle.com/c/lehighcse347447